iio: buffer-dmaengine: generalize requesting DMA channel
This patch generalizes the iio_dmaengine_buffer_setup_ext() functions
by passing the pointer to the DMA channel as an argument rather than
the channel name. This will allow future callers of the function to
use other methods to get the DMA channel pointer.

This aims to keep it as easy to use as possible by stealing ownership
of the dma_chan pointer from the caller. This way, dma_request_chan()
can be called inline in the function call without any extra error
handling.

Signed-off-by: David Lechner <dlechner@baylibre.com>

v3 changes:
* This is a new patch in v3.