arm: dts: make use of AXI DMA bus type macros
Replace use of hardcoded values with the AXI DMA bus type macros.

There is already a dt-bindings/dma/axi-dmac.h file that defines macros
for the adi,source-bus-type and adi,destination-bus-type properties.
However, this was only being used in about half of the dts files that
could make use of it. Update the rest of the dts files to use these
macros to make it easier to see what is actually being specified.

Signed-off-by: David Lechner <dlechner@baylibre.com>